// src/app/interfaces/datos-pdf.interface.ts
export interface DatosPdf {
  contenidoBase64: string;
  extension: string;
  fechaModificacion: string;
  fechaSubida: string;
  idVersion: number;
  nombreArchivo: string;
  tamañoBytes: number;
  tamañoKB: number;
  tamañoMB: number;
  tipoMIME: string;
  ultimaModificacion: number;
  usuario: string;
}